requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.15.0
websockets>=12.0
